Move to_struct, getitem tests
