+++ python/cudf/cudf/tests/series/accessors/test_struct.py
@@ -11,3 +11,25 @@ def test_struct_iterate_error():
     )
     with pytest.raises(TypeError):
         iter(s.struct)
+
+
+@pytest.mark.parametrize(
+    "data",
+    [
+        [{}],
+        [{"a": None}],
+        [{"a": 1}],
+        [{"a": "one"}],
+        [{"a": 1}, {"a": 2}],
+        [{"a": 1, "b": "one"}, {"a": 2, "b": "two"}],
+        [{"b": "two", "a": None}, None, {"a": "one", "b": "two"}],
+    ],
+)
+def test_struct_field_errors(data):
+    got = cudf.Series(data)
+
+    with pytest.raises(KeyError):
+        got.struct.field("notWithinFields")
+
+    with pytest.raises(IndexError):
+        got.struct.field(100)
