+++ python/cudf/cudf/tests/test_struct.py
@@ -7,7 +7,6 @@ import pytest
 
 import cudf
 from cudf.testing import assert_eq
-from cudf.testing._utils import DATETIME_TYPES, TIMEDELTA_TYPES
 
 
 @pytest.mark.parametrize(
@@ -196,102 +195,3 @@ def test_dataframe_to_struct():
     with pytest.warns(UserWarning, match="will be casted"):
         got = df.to_struct()
     assert_eq(got, expect)
-
-
-@pytest.mark.parametrize(
-    "series, slce",
-    [
-        (
-            [
-                {"a": "Hello world", "b": []},
-                {"a": "CUDF", "b": [1, 2, 3], "c": 1},
-                {},
-                None,
-            ],
-            slice(1, None),
-        ),
-        (
-            [
-                {"a": "Hello world", "b": []},
-                {"a": "CUDF", "b": [1, 2, 3], "c": 1},
-                {},
-                None,
-                {"d": ["Hello", "rapids"]},
-                None,
-                cudf.NA,
-            ],
-            slice(1, 5),
-        ),
-        (
-            [
-                {"a": "Hello world", "b": []},
-                {"a": "CUDF", "b": [1, 2, 3], "c": 1},
-                {},
-                None,
-                {"c": 5},
-                None,
-                cudf.NA,
-            ],
-            slice(None, 4),
-        ),
-        ([{"a": {"b": 42, "c": -1}}, {"a": {"b": 0, "c": None}}], slice(0, 1)),
-    ],
-)
-def test_struct_slice(series, slce):
-    got = cudf.Series(series)[slce]
-    expected = cudf.Series(series[slce])
-    assert got.to_arrow() == expected.to_arrow()
-
-
-def test_struct_slice_nested_struct():
-    data = [
-        {"a": {"b": 42, "c": "abc"}},
-        {"a": {"b": 42, "c": "hello world"}},
-    ]
-
-    got = cudf.Series(data)[0:1]
-    expect = cudf.Series(data[0:1])
-    assert got.to_arrow() == expect.to_arrow()
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [{}],
-        [{"a": None}],
-        [{"a": 1}],
-        [{"a": "one"}],
-        [{"a": 1}, {"a": 2}],
-        [{"a": 1, "b": "one"}, {"a": 2, "b": "two"}],
-        [{"b": "two", "a": None}, None, {"a": "one", "b": "two"}],
-    ],
-)
-def test_struct_field_errors(data):
-    got = cudf.Series(data)
-
-    with pytest.raises(KeyError):
-        got.struct.field("notWithinFields")
-
-    with pytest.raises(IndexError):
-        got.struct.field(100)
-
-
-@pytest.mark.parametrize("dtype", DATETIME_TYPES + TIMEDELTA_TYPES)
-def test_struct_with_datetime_and_timedelta(dtype):
-    df = cudf.DataFrame(
-        {
-            "a": [12, 232, 2334],
-            "datetime": cudf.Series([23432, 3432423, 324324], dtype=dtype),
-        }
-    )
-    series = df.to_struct()
-    a_array = np.array([12, 232, 2334])
-    datetime_array = np.array([23432, 3432423, 324324]).astype(dtype)
-
-    actual = series.to_pandas()
-    values_list = []
-    for i, val in enumerate(a_array):
-        values_list.append({"a": val, "datetime": datetime_array[i]})
-
-    expected = pd.Series(values_list)
-    assert_eq(expected, actual)
